requirements.txt
Pillow==8.1.0
selenium==3.141.0
